updated everything to match the usage of 32-bit unsigned by digest now. Also realized that threading won't really work if the string has non-actg characters in it

+++ src/thread_out.cpp
@@ -4,7 +4,7 @@ namespace thread_out
 {
 
 void thread_mod(unsigned thread_count, std::vector<std::vector<size_t>>& vec, 
-    const char* seq, size_t len, unsigned k, uint64_t mod, uint64_t congruence, size_t start, 
+    const char* seq, size_t len, unsigned k, uint32_t mod, uint32_t congruence, size_t start, 
     digest::MinimizedHashType minimized_h){
         int num_kmers = (int)len - (int)start - (int)k + 1;
         if(k < 4 ||start >= len || num_kmers < 0 || (unsigned)num_kmers < thread_count){
@@ -37,13 +37,13 @@ void thread_mod(unsigned thread_count, std::vector<std::vector<size_t>>& vec,
     }
 
 void thread_mod(unsigned thread_count, std::vector<std::vector<size_t>>& vec, 
-    const std::string& seq, unsigned k, uint64_t mod, uint64_t congruence, size_t start, 
+    const std::string& seq, unsigned k, uint32_t mod, uint32_t congruence, size_t start, 
     digest::MinimizedHashType minimized_h){
         thread_mod(thread_count, vec, seq.c_str(), seq.size(), k, mod, congruence, start, minimized_h);
     }
 
 void thread_mod_roll(std::vector<size_t>& vec, const char* seq, 
-    size_t ind, unsigned k, uint64_t mod, uint64_t congruence, 
+    size_t ind, unsigned k, uint32_t mod, uint32_t congruence, 
     digest::MinimizedHashType minimized_h, unsigned assigned_kmer_am){
         digest::ModMin dig(seq, ind + assigned_kmer_am + k -1, k, mod, congruence, ind, minimized_h);
         dig.roll_minimizer(assigned_kmer_am, vec);
